day_05/requirement.txt
asgiref
autopep8
Django
psycopg2-binary
pycodestyle
pytz
sqlparse
toml